+++ api/Ourchitecture.Api/Program.cs
@@ -15,23 +15,12 @@ namespace Ourchitecture.Api
         {
             HostFactory.Run(x =>
             {
-                //var init = new Startup();
-           
-                //x.UseLinuxIfAvailable();
                 x.Service<NancySelfHost>(s =>
                 {
                     s.ConstructUsing(name => new NancySelfHost());
 
-                    s.WhenStarted(tc =>
-                    {
-                        tc.Start();
-                        //init.InitializeRhino();
-                    });
-                    s.WhenStopped(tc =>
-                    {
-                        tc.Stop();
-                        //init.CleanupRhino();
-                    });
+                    s.WhenStarted(tc => { tc.Start(); });
+                    s.WhenStopped(tc => { tc.Stop(); });
                 });
 
                 x.RunAsLocalSystem();
@@ -49,7 +38,7 @@ namespace Ourchitecture.Api
 
         public void Start()
         {
-            RhinoLib.LaunchInProcess(RhinoLib.LoadMode.Headless, 0);
+            Startup.LaunchInProcess(Startup.LoadMode.Headless, 0);
             Console.WriteLine("Rhino loaded at port 88.");
             m_nancyHost = new NancyHost(new Uri("http://localhost:88"));
             m_nancyHost.Start();
@@ -63,21 +52,4 @@ namespace Ourchitecture.Api
             Console.WriteLine("Stopped. Good bye!");
         }
     }
-
-    /*
-    public class Program
-    {
-        public static void Main(string[] args)
-        {
-            var host = new WebHostBuilder()
-                .UseContentRoot(Directory.GetCurrentDirectory())
-                .UseKestrel()
-                .UseStartup<Startup>()
-                .UseUrls("http://localhost:88/")
-                .Build();
-
-            host.Run();
-        }
-    }
-    */
 }
